Fix clang-tidy non-stylistic issues: remove dead code, add missing default cases.
- Hunt2.cpp: Remove dead store of appRunning, needMenus, and b variables
- Characters.cpp: Remove dead store of v variable, add default case to switch
- Audio_SDL.cpp: Make implicit widening conversions explicit with static_cast
- CustomMaterials.cpp: Add default case to switch statement
- Game.cpp: Remove dead stores of r variable, add default cases to 3 switch statements

+++ Audio_SDL.cpp
@@ -145,10 +145,10 @@ static void SDLCALL AudioCB(void* /*userdata*/, Uint8* stream, int len)
     // ── Clamp and write stereo output ──
     short* out = (short*)stream;
     for (int i = 0; i < n; i++) {
-        out[i * 2]     = (short)std::max(-32768, std::min(32767, L[i]));
-        out[i * 2 + 1] = (short)std::max(-32768, std::min(32767, R[i]));
+        out[static_cast<size_t>(i) * 2]     = (short)std::max(-32768, std::min(32767, L[i]));
+        out[static_cast<size_t>(i) * 2 + 1] = (short)std::max(-32768, std::min(32767, R[i]));
     }
-    if (n < nSamples) memset(out + n * 2, 0, (nSamples - n) * 4);
+    if (n < nSamples) memset(out + static_cast<size_t>(n) * 2, 0, static_cast<size_t>(nSamples - n) * 4);
 }
 
 // ─── Public API ───────────────────────────────────────────────────────────────
